requirements.txt
Flask==2.1.1
Flask-WTF==1.0.0
tensorflow==2.7.0
tensorflow-hub==0.12.0
Pillow==9.0.1
numpy==1.21.3
protobuf==3.20.0
Werkzeug==2.1.1